+++ packages/atproto/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.5.8
 
 - Added `cursor` to `Repos` object. ([#627](https://github.com/myConsciousness/atproto.dart/issues/627))
+- Upgraded `atproto_core` and exposed `RateLimit` and `RateLimitPolicy` objects. ([#636](https://github.com/myConsciousness/atproto.dart/issues/636))
 
 ## v0.5.7